+++ R/ml_fit_entropy_o.R
@@ -12,7 +12,7 @@
 #' @export
 #' @examples
 #' path <- system.file("extdata/minitoy", package="MultiLevelIPF")
-#' ml_fit(ref_sample = import_IPAF_results(path))
+#' ml_fit_entropy_o(ref_sample = import_IPAF_results(path))
 ml_fit_entropy_o <- function(ref_sample, controls, field_names,
                              BBsolve_args = list()) {
   .patch_ml_fit_args()