+++ R/MultiLevelIPF-package.R
@@ -8,7 +8,7 @@
 #' \tabular{ll}{
 #'   Package: \tab MultiLevelIPF\cr
 #'   Type: \tab Package\cr
-#'   Version: \tab 0.0.7\cr
+#'   Version: \tab 0.0.8\cr
 #'   Date: \tab 2014-06-17\cr
 #'   License: \tab GPL v3\cr
 #' } ~~